fix(gateway): Fix JSON Scalar
The worker response relied upon a mixture of custom String formatting
and the Display implementation of ConstValue which led to an incorrect
serialization of our custom JSON scalar.

Instead of formatting, we create a `serde_json::Value` which avoids any
formatting issues.

Unfortunately, we lose the ordering of the fields requested by the
query/mutation. I tried using `preserve_order` feature on serde_json but
it breaks a lot of tests. The best solution was to make them stable
with the feature `redactions` of insta. But, for some reason it did work
when doing `cargo test -p <package>` but not with
`cargo test --workspace --exclude e2e-tests`. So for now,
I gave up on this.