fix(gateway/dynamodb): Fix nested relation creation
Adding the NewType `CurrentDateTime` to have a consistent
`Utc::now()` for the whole request. This has the nice benefit
that all createdAt/updatedAt will be the same for all created/updated
nodes during a mutation.

When creating a relation we keep the existing createdAt/updatedAt.
The added relation row is expected to be copy of the "real" row so it
shouldn't have different dates.

Fixes: GB-2751,GB-2541,GB-2743