A less repetetive naming convention for relation inputs
The old naming convention for relation inputs resulted in quite
repetetive names like `UserUserRelatePostPostCreateInput`, as each type
was prefixed by the name of both models involved and the name of the
relation (which by default also has the name of both models involved).

I've updated the prefix to just use the relation name, and have put the
model being created/updated as part of the input into a later part of
the name, so you end up with `UserRelatePostCreatePostInput` instead.

It's less repetetive and I think it should still be unique enough as a
given model should only have one relationship with a particular name.

Let me know if there's something I've not thought of or if there's any
better ideas for naming conventions.