Add support for unique indexes on multiple fields.
Prior to this change `@unique` could be specified on a field but there
was no way to have a unique index on multiple fields.

This updates `@unique` with an optional `fields` parameter which is a
list of additional fields that should be included in the index.  These
will be hashed as part of the ConstraintID.

These multi-field indexes can also be looked up in `by: ` fields: for
example with

```graphql
type Post {
  id: ID!
  user: User!
  slug: String! @unique(fields: ["user"])
}
```

You can query:

```graphql
query {
  post(by: {slugUser: {slug: "my-cool-post", user: "not-even-sure-is-this-possible?") {
    id
  }
}
```