feature(search): Add search lambdas
Adding 3 lambdas for search:
- deploy: responsible to update the indices for a DynamoDB
  table according the new schema.
- indexer: Update the indices from DynamoDB streams.
- query: Executes a query on the indices.

All rely on a new package common/search which provides a generic
interface to interact with Tantivy indices (not tied to DynamoDB).

Schema udates are really simple, on any change, we just re-create a new
index.

Write performance isn't optimized and will certainly be the bottleneck,
but with rough lambda batching it should hold for now.

Read queries are for now expected to run under 200ms as higher than that,
it's not guaranteed that the underlying files won't be deleted. Currently
the writer just waits for that dealy before deleting any files. Later this
should be done asynchronously through a SQS queue, allowing us to increase
that delay.

Resolves: GB-3122