First pass at OpenAPI parsing
This commit adds a `parser-openapi` library to the repo that can generate a
GraphQL schema for any GET operations in an OpenAPI schema.  It gathers up all
of the GET operations and adds them to the query type, and also outputs the
datatypes that will be returned.

Currently it just outputs some schema SDL, as that was easiest for testing.  A
future PR will integrate it into the rest of the grafbase codebase somehow
(either by feeding the SDL into the existing parser, or by manipulating the
Registry directly, not quite sure yet).

There's still a fair bit to do on this, currently it doesn't handle any of the
mutations, doesn't handle any parameters and the shape of the schema is quite
far from ideal at the moment.But it's a starting point that we can build on
(probably after getting some of the actual HTTP code in place).

### Implementation notes

There's a bunch of differences between OpenAPI & GraphQL that makes mapping
direct from an OpenAPI spec to a GraphQL schema challenging.

1. OpenAPI uses a single "schema" type to represent pretty much everything in a
GraphQL spec: Objects, Input Objects, Unions, Scalars, Lists etc.
2. It's very easy to end up with un-named objects or unions in an OpenAPI spec.
3. Schemas can use "references" to link in to other schemas (which can be
basically any kind of thing that appears in GraphQL).
4. Schemas are roughly based on JSON schema which is more focussed on
validating data than it is describing its shape (e.g. you can say some data
should _not_ match a particular schema)

I've implemented the parser by taking the OpenAPI spec and building up a
slightly transformed graph of the various schemas/responses, and then using
that to build up the output.  This made it slightly easier to deal with all of
the above, as all the references are now just edges in the graph, the shape of
the graph is closer to GraphQL semantics, and we can easily traverse the graph
in both directions (e.g. when trying to name something based on its ancestors).