Parse the openapi directive
This commit adds parsing of the `@openapi` directive to the parser.
This gets returned in the `ParseResult` which will be used to fetch and
parse the OpenAPI specs in my next PR.

I've taken a fairly different approach to parsing this directive: rather
than taking the `ConstDirective` and parsing it manually I've
implemented a `serde::Deserializer` that understands `ConstDirective`.
This means we can just derive `Deserialize` on a type and we get parsing
more or less for free.

Not sure if the error messages are going to be quite as good mind you -
serde is pretty terse by default, and telling what bit of the document
it's referring to is tricky without some spans.  But I think we could
probably get spans on the individual `ConstValue`s inside the directive
and then use them to fix the problem.  But that's a task for a
different day.