Support OpenAPI path parameters
openapi has a concept of path parameters - parameters that go into the paths of
a url.  this pr adds support for scalar path parameters.  the openapi spec also
supports lists and objects in path position, along with approximately 6
different ways to encode them (openapi is a very annoying spec), but i think
these are relatively uncommon so for now i'm not bothering with them.

Currently these parameters just appear as top-level arguments on their
respective fields.  I reckon at some point we'll want to consolidate these into
a single `entity` or `entityCollection` field, and these will probably end up
properties on an `EntityBy` type, but that's work for the future.