Handle some common QueryParameter encoding styles
OpenAPI query parameters are modelled with OpenAPI schemas but the
actual representation of them is a series of keys & values.  OpenAPI
defines 8 different ways to encode from lists & objects into actual
query parameters (based on the `style` & `explode` parameters).

This PR adds support for three of the more common styles: form exploded
(the default), form unexploded and deep object (the one that stripe
uses).

I've also added some validation on the schemas of parameters - it's not
clear how nested lists or objects would work under form encoding (and
the spec doesn't mention them) so for now we'll error if we get a schema
that looks like this.