Parse mutations out of OpenAPI specification.
Prior changes handled taking the HTTP GET operations out of an OpenAPI spec and
generating queries based on them.  This PR handles taking POST/PUT/DELETE/etc
operations and making mutations out of them.

This involved a bit of a re-organistion of the code.  I'd made some bad choices
initially and crammed far too much into the `query_operations` module.  I've
now split it up into `parameters` and `operations`.

This also adds a test of parsing the OpenAPI petstore spec, as the stripe one
is just to unweildy to tell whether it's doing the exact right thing.

This PR still doesn't contain the implementation of mutations - that's
up next.