Re-case OpenAPI enums to SCREAMING_SNAKE_CASE.
GraphQL enums are conventionally SCREAMING_SNAKE_CASE, but OpenAPI enums
are not.  I didn't bother handling this when I first added enums, but
now I'm coming back to it.

Re-casing them in our schema is pretty straight forward but we also need
to handle sending the correct case on to the downstream API.  The
easiest way to facilitate this was putting a `value` field into the
`MetaEnumValue` type that contains the remote value.

For re-casing the return value I've just hooked into the existing
resolvers: for enum fields we first resolve the field out of its parent
object then run it through a `RemoteEnum` resolver that looks up the
`MetaEnumValue`s and converts the input.

For re-casing the inputs I've hooked into the `resolve_input` function.

I'm not certain I've picked the best way to do this, so feedback
welcome.