GB-3190: Search query & filter API
Search query & filters are processed in three stages:

1. `dynaql` parses the actual query into a `grafbase_runtime::search::GraphqlQuery`.
2. In the worker, it's transformed into a `search::Query` which is close to Tantivy's
   query model but easier to manipulate.
3. In the lambda, we build the Tantivy query from it with `search::TantivyQueryBuilder`.

The last two are unit-tested, the first was only checked through the CLI manually. While
I have a working CLI implementation I haven't updated the tests there yet.

I've also refactored various things during this commit and moved code in the search lambdas.
The application logic is now in a lib crate as it's supposed to provide lower compilation
times. The `search-dyanmodb-indexer` tests are now in a separate file. And the `search-query`
lambda was renamed `search-query-executor`.

Resolves: GB-3190