Check expected status codes in HTTP resolver
OpenAPI provides us with the status codes that we should expect for each
response.  I hadn't bothered implementing any checking of the status code so
far, so when we got an error back we would just pass whatever JSON the remote
API returned.  This would usually not conform to the schema we'd defined and
the user would receive a fairly useless error message.

This PR updates the code to check status before decoding JSON.   If the status
is not as expected we'll return an error explaining that's what happened.