Fix a parser crash when users use underscores in model names.
When a user gives us a type with an underscore in it we convert it to
PascalCase prior to using it.  There was a bit of the parser where we didn't
use this converted name, which was leading to us having both `PKS_Options` and
`PKSOptions` in `schema_to_build` even though only one of those types actually
existed.