Don't panic if users schema has clashing types
We weren't correctly setting the rust_typename on an InputObject we
generated, which causes the parser to panic if the user tries to create
a type thats name clashes with the generated type.

This fixes the panic, but I'm not sure the user is going to get a sane
schema if they try and do this, so we might want to do some more work
around error detection at some point?  Not sure.