Interpolate env vars into all the @openapi parameters
Fredrik was trying to pull in some tiny bird OpenAPI stuff and wanted
to use env vars to populate the schema & API URLs.  I hadn't explicitly
used the DynamicString type for these fields, so that wasn't possible.

I started off updating the fields to use DynamicString, but it was kind
of a pain. I was using serde to parse out `Url` types.  I converted that
to a `DynamicString` but then would have to explicitly evalute those
strings, _then_ do some `Url` parsing by hand.

I also couldn't just export the `OpenApiDirective` type from
`parser-openapi` as it no longer represented the final result of
parsing, which nearly led me down the route of having two type
heirarchies.  Yuck.

This PR takes a different approach: instead of using the DynamicString
directly to determine where to interpolate, it hooks into the
`ValueDeserializer` I'd already written and interpolates any type that
deserializes as a string.

Makes it much easier to use for development, and also makes it harder to
forget to add env var interpolation to a directive.  This might require
some adjustment if we ever need to support runtime variables, but for
now it works quite nicely.