Optimise introspection queries.
When using `@openapi` on a large spec (e.g. stripe) introspection queries take
an absolute age (about 8 seconds) and consume a ton of memory (enough to OOM
prod, in the CLI `node` is using about 600MB by the time it's done).  Obviously
this ain't great.

Looking into it: this seems to be an artifact of the mismatch between old
`async_graphql` code and new dynaql stuff.  Newer dynaql stuff puts it's data
into a graph, the `async_graphql` stuff wants to return a `Value`.  Most of our
code works in the new `dynaql` world, but introspection is still based on
`async_graphql` code.

The glue between these ends up putting things into the response graph, then
immediately generating a `Value` out of the data that was just put into the
graph, which then gets put back onto the graph, then converted back into a
`Value` etc. etc. for each resolve call it needs to introspect.

To make things worse, it also wasn't removing the intermediate values from the
graph when converting to a `Value`.  In my testing this meant the graph had
just shy of a million entries, and consumed somewhere upwards of 100MB just for
the book-keeping, not the actual data.  EEK.

I tried removing the values from the graph as part of the conversion back to a
`Value` which helped memory significantly.  But unfortuanately it was still
slow af, so I've ended up adjusting all the `async_graphql` stuff to put things
straight into the graph.

Introspection now takes around 1 second, and `node` in the CLI is sitting at
240MB instead of 600MB after introspection.  I am worried that this will still
be too much memory on production, but want to try it out and see...