feature(gateway): Add account_id in response headers
For usage and billing we need usage events to be associated with an account.
The project_id was already there and we could have joined on TinyBird side,
but it opens the door to sync issues between TinyBird and prod for projects/accounts.
And it's just easier than managing a streaming join as we want usage events to be available
as soon as possible.

Resolves: GB-3465