Fetch & parse @openapi prior to parsing the main schema.
We've got a customer that wants to use an `@openapi` generated type in their
schema with a custom resolver that they have written.  Currently we parse the
openapi schemas after the entire main parsing process has finished, so this
errors out because the main parser has no idea about the connector generated
types.

The easiest way to handle this use case is to handle the openapi specs prior to
parsing the main specification.  This is a bit awkward with the current code
because we need the main parser to get at the directive that tells us about the
connectors.

So, I've updated the main parser to call to adopt a slightly different flow:

1. It parses out the directives.
2. It passes control over to `parser-openapi` to do its thing (via a trait so
   we can mock in tests and avoid circular dependencies).
3. We continue with the rest of the parsing process.

This should be adaptable to other connectors we want to build as well.