Specify queryNaming strategy in openai template
By default the OpenAPI connector uses schema names to define the names of the query & mutation fields.  

But the specification we're using for the OpenAI example is structured to use responses as it's schemas rather than resources.  This means that by default we get rather poor query & mutation names for OpenAI.

This PR updates the OpenAI template to specify the OPERATION_ID naming strategy, which improves things somewhat.