Parse input objects into the registry
Prior to resolvers we had no reason to pull input objects from the
source schema into the registry, as there was no way to do anything with
that custom type.

Resolvers has changed that though - now that we support them we need to
allow users to provide custom input types, which this PR does.