Fix parent data on resolvers when queried by id
Some resolvers need data from their containing type to function
correctly.  We added a hack to provide this data prior to releasing
resolvers, but it wasn't working in every case.

It relied on the assumption that the first entry in the
`last_resolved_value` object was the parent object, but it seems like
some resolvers populate the `last_resolved_value` with relation data as
well, and in that case this was falling over.

I've updated it to walk up the resolver_nodes to find what's _hopefully_
the type name and use that key instead, but to fallback to the previous
behaviour if that doesn't work.  Also added an e2e test of this specific
case.

Imagine there might be situations where this doesn't work as well...?
But given this is a temporary hack till query planning is in place it's
probably fine?