Calculate discriminators for OpenAPI unions.
OpenAPI schemas have a bunch of ways of doing polymorphism: oneOf,
allOf, anyOf, etc - we map most of these into GraphQL union types.

OpenAPI also has the concept of a "discriminator" - a field that you can
use on one of these polymorphic types to determine which variant you've
actually got when you're looking at the data.  Unfortunately for us,
almost nobody uses this functionality.  Very helpful.

Since GraphQL needs to know precisely which variant we have, we'll need
to figure this out for ourselves.  The "correct" way to do this is
probably to use the JSON schema from the OpenAPI spec to validate each
variant in turn and pick the first one that passes.  But, that'd involve
pulling in an absolute shitload of JSON schema to our already far too
big registry.

So, I've implmeneted a bit of a hack that looks through each of the
types in a union and plays a game of spot the difference, recording that
in a `UnionDiscriminator` type in the registry.

I've also added a resolver that takes this, applies it to some data, and
then populates that data with a `__typename` accordingly.

This won't work on its own, but my next PR updates the dynaql resolution
mechanism such that it does.