Handle unions that have individual scalar varaints.
Stripes OpenAPI spec has quite a lot of "expandable" fields.  The schema
for these tends to be `string | SomeObject` - where the API will return
the string ID of the entity by default, but will expand if you pass in
some parameters.

We weren't handling the scalar variants when generating our unions, as
unions can't contain scalars in GraphQL.  This changes that - when we
see one of these variants we wrap the scalar in a type that looks like:

```graphql
type StripeString {
   data: String
}
```

This allows these types to be represented in GraphQL, so that fields in
stripe that follow this pattern aren't completely broken.