Add a test of `@openapi` unions
This adds a test of the `@openapi` union fixes I've been working on over
the past week and a bit.  Quite hard to test these thoroughly in the
`api` repo right now (although I want to work on that), so at least
wanted to get a test in this repo in the meantime.