Split graph construction out of openapiv3 parsing.
I'm about to implement parsing of openapi v2 specifications.  I'd
initially hoped that I could do this by restructuring the data in memory
and just passing it through the v2 parsing code. But having made a start
on that it seems unlikely to work well: the structures are different
enough that it'd be a bit of a lossy conversion and wouldn't output a
particularly nice schema.

So, I've decided to write separate parsing of the v2 specification into
our intermediate graph structure and then keep the output code roughly
the same.  This commit is prep for that: it splits some graph
construction code out of the v3 parsing so it can be shared, and makes a
handful of other hopefully useful changes.

I've also backported some changes from #2119 to try and get them in
without the issues it's causing - they're definitely an improvement so
lets see.

Fixes GB-3601 (part of it anyway - just putting this here so it gets
linked in linear)