add conversion from `cynic_introspection::Schema` to `Registry`
The conversion takes all the known information from the introspected
schema and adds it to the registry.

The conversion is infallible (using `From` instead of `TryFrom`), but is
allowed to panic, if a future update to `cynic_introspection` introduces
new enum variants to any of the non-exhaustive enums in the crate. This
_should_ not happen, unless the actual GraphQL specification
(https://spec.graphql.org) changes.

Signed-off-by: Jean Mertz <git@jeanmertz.com>