add test-case for introspection-to-registry conversion
A snapshot of the publicly available Star Wars GraphQL API schema
(https://swapi-graphql.netlify.app) is stored locally (to avoid failing
tests when the upstream API changes).

The actual test results in a 5000+ lines snapshot. This was validated
manually once, and will help us in the future, when changes are made to
the conversion logic (unlikely).

Signed-off-by: Jean Mertz <git@jeanmertz.com>