add stubbed GraphQL resolver implementation
This resolver allows querying upstream GraphQL APIs, passing along the
returned data to the caller.

This commit does *not yet* implement the serializer to sent the correct
query or mutation to the upstream server, but it does add all the other
plumbing.

Similarly, it does not yet hook into the Grafbase schema, so `@query`
directives aren't in yet, and there's no way to trigger this actual
logic, but that too will be added in a follow-up commit.

Signed-off-by: Jean Mertz <git@jeanmertz.com>