Fix schemas that use `JSON` & `@openapi`
`@openapi` declared a JSON type, which we were naively converting into a
`TypeDefinition` for an `Object`.  This confused the rest of the parser
as it tried to treat a `JSON` field like it was an `Object`.

I've fixed this by being a bit more (though not completely) thorough
when reverse engineering TypeDefinitions.

I've also removed the JSON scalar definition from `parser-openapi` as I
don't think it's needed with the way we currently integrate into the
parser - it was mostly there to support the way we used to do things.