Preserve enum defaults in registry JSON.
Fredrik was testing tinybird with `@openapi` and noticed that the
default values for the `format` argument weren't being set properly.

This turned out to be because this argument is an enum.  We store & load
the registry as JSON, and JSON has no syntax for enums.  So any enum
default value would get converted to a String, and then all the tooling
would ignore it because it expected an enum.

To fix, I've added custom `Serialize` & `Deserialize` impls to the
`default_value` field that'll store enums as `{"__enum": "VALUE"}`.

Quite a lot of work for a relatively simple problem but at least it
works.