Add introspectionHeaders to `@openapi`
Some OpenAPI specifications might be behind auth so you need to send
some headers to get at them.  This PR adds that functionality to
`@openapi` under a new parameter `introspectionHeaders`.

Didn't want to just reuse the headers property because some
specifications aren't hosted in the same place as the API (e.g. Stripe
is on GitHub and we wouldn't want to send API keys to GitHub)