Make the `@openapi` url parameter optional.
OpenAPI has a servers entry that is often populated with the URL of the
API.  In these cases there's not much point in providing `url` to
`@openapi` - so I've made it optional for those cases.

There is an edge case where servers is a relative URL, but the only
example I've seen of this is petstore.  Most of the actual APIs we're
testing with provide the full URL.  For those cases I'm still expecting
users to provide a `url` to `@openapi`