handle partial GraphQL responses
As specified in the GraphQL spec[1], field errors allow for partial
return values, while returning `null` for fields that could not be
resolved.

To support this, we have to return whatever `data` we get back from the
upstream server, regardless if it returned an error.

This does expect the upstream server to adhere to the GraphQL spec, but
that's an assumptions we've been making all along.

[1]: https://spec.graphql.org/October2021/#sec-Handling-Field-Errors

Signed-off-by: Jean Mertz <git@jeanmertz.com>