pass variable values to upstream GraphQL server
The serializer does not resolve variables, instead serializing variable
references as-is (e.g. `$foo`).

This commit ensures that the constant values these references point to
are embedded in the HTTP request to the upstream server.

Only variable values actually referenced in the query or mutation are
embedded, any local-only variables are filtered out for security and
privacy reasons.

Signed-off-by: Jean Mertz <git@jeanmertz.com>