Don't set entity_type on sqlite Record for constraints.
PR #2155 added entity_type into a couple of places, but I don't think it
should have been added here.  Adding it to the Record causes the
constriant to be returned in get by ID queries, which throws errors in
the resolution mechanism.

This causes the CLI tests to fail.

I'm assuming (but would be good for @miaxos or anyone else in the know
to confirm) that it's enough for it to be in the document but not the
sqlite Record itself.