add `PassThrough` parsing mode for scalar values
When parsing values from an upstream GraphQL server, we can expect
custom scalar types to be part of the schema. Up until now, however, the
code assumed all scalar types are known, and were thus hard-coded.

To mitigate this, we now allow defining the type of scalar parser to use
when deserializing the scalar value.

To start, we support two types of parsers, the `BestEffort` parser is
the default, backward-compatible, parser that parses scalar values based
on a hard-coded list of potential types, and errors if none of them
match.

The new `PassThrough` parser parses the value into `serde_json::Value`,
and leaves it at that.

This solution works for both the existing schemas out there, and new
ones that embed upstream GraphQL APIs. More parsers could be added in
the future, but are currently not needed.

Signed-off-by: Jean Mertz <git@jeanmertz.com>