Pull in tar-rs fork to work around startup issue.
`tar-rs` has an issue where it won't overwrite files if they're stored
as hard links in the tar.  We install our asset `node_modules` with
pnpm, which uses hard links to deduplicate content.  This means
`grafbase dev` will refuse to overwrite some existing files on startup.

I've fixed this PRd it, this pulls the fix in the meantime.