chore: Clean up some dynaql query planning code (#2251)
I've been trying to get to grips with query planning. I'll be making PRs
to rename/clarify/simplify/document things as I go along, which will
help me (and hopefully others?) understand things a bit better.

First up: there's a module in `dynaql` named `logical_plan_utils`. It
clearly started life as a copy & paste of `resolver_utils`. This is an
attempt to clean it up a bit:

1. Firstly, I renamed the module to `planning` - any module with `utils`
in the name is destined to become an unclear dumping ground imo.
2. Moved a lot of the code from `container` to `mod.rs` because it
looked like it wasn't doing much (this probably made the diff way worse
than it is, sorry about that).
3. Renamed a lot of the functions from `resolve_xyz` - we're building a
plan not resolving so the old names didn't make much sense.
4. I moved a function out of the registry into `planning` that really
had no reason to be there.
5. Tried to clarify a few comments

There shouldn't really be any logic changes here, so I wouldn't dig too
deep into the code for this PR