fix(graphql connector): don't error on missing nullable object (#2260)
Previously, if the GraphQL connector returned `null` for a nullable
object, the resolver would still try to resolve the embedded fields
within the (missing) object, which would also return `null`, causing an
error for any of those fields which were non-nullable.

This was caused by not setting `early_return_null` to `true` for objects
which were null.

The solution to this, was to switch from using a resolver (specifically,
`ResolverType::ContextDataResolver`) to using a transform
(`Transformer::JSONSelect`). The transform has the check for null values
and auto-sets the `early_return_null` flag, while the resolver does not.

It _appears_ this is by design, and we do not actually want the resolver
to set this flag on `null` values, but that investigation was
inconclusive.

In any case, switching to the transform fixes the issue as expected.
Adding a test for this turned out to be more cumbersome than expected,
but will be added in a follow-up commit.

While fixing the above described issue, another related issue was fixed
wherein either the `http` or `graphql` connector could return `null` as
the final resolved value, without setting `early_return_null`. Hitting
this case (especially for the `http` connector) is unlikely, but either
way, this has been fixed as well.

Signed-off-by: Jean Mertz <git@jeanmertz.com>