fix: allow connector types to be extended (#2289)
Several of us on were convinced that we allowed users to extend
connector types with additional resolver fields. On investigation this
turned out not to be the case: we'd added support for using connector
types in resolvers on user owned types, but not the other way round.

This is a feature we want to support though, so this PR fixes it.