fix: internal error on unlinking entities in an update (#2306)
Authdog were running into "internal errors" when they tried to run an
update that unlinked an entity. This was pretty trivial to reproduce,
and seemed to be because we were emitting an `UpdateNodeInternalInput`
_and_ a `DeleteAllRelationsInternalInput` for the edge in the database.
This ran into issues because both these operations want to act on the
same row of the database, but dynamo transactions only allow one
transaction item per database and our graph_transaction code doesn't
know how to consolidate them.

I've fixed this by instead emitting an `UpdateRelationInternalInput` for
the edge row, and then updating the deduplication code to merge an
update & delete for the same relation row into just a delete.

I'm not entirely sure if this is correct, since very little is
documented and apparently there's also very few tests of this. But I've
added a couple of tests: one for this specific case and one for a
slightly different relation update scenario, and they both seem to be
fine.

CLI version of these tests is [over
here](https://github.com/grafbase/grafbase/pull/439)

Fixes GB-3892