fix: support queries with recursion (#2318)
There's a long standing bug in the gateway where if you try to make a
query that recurses it won't work (think it ends up hanging/crashing
with a stack overflow, but I'm not certain). Similarly if you make a
query that doesn't recurse but does end up with the same database entry
in multiple places in the response you'll get back a malformed response
with only one instance of that entry.

It's _really_ easy to encounter this so I thought I should fix before a
user notices.

The underlying reason is the `QueryResponse` structure we put things
into likes to index database items by their ID, and assumes that an item
can only appear at one point in the response. It's a pretty silly
assumption. I am guessing the idea was for this structure to model the
database in memory, which could then be used to build a response - but
that is quite far from what it's actually doing.

To fix I've updated QueryResponse to support database entries appearing
in more than one place, and removed _most_ of the logic that handled
replacing entries (apart from some live query support stuff).

Fixes GB-2538
Fixes GB-3900