chore(graphql connector): provide stable ID for each connector (#2324)
This change allows the connector to know its stable ID within a given
schema.

This will be used in the future to store connector-specific HTTP headers
in the global `Registry`, without the need for the user to manually
provide a stable ID (such as the connector name, as that name will be
become optional in the future).

The change is deliberately scoped as narrowly as possible (e.g. only for
the GraphQL connector, and not encoded in the `parser::ConnectorParsers`
trait). Once the optional namespacing feature lands for the GraphQL
connector (a feature specifically needed by one of our customers), it
will be extended to the OpenAPI as well.