enhance(graphql connector): allow unnamed connectors (#2325)
This change allows GraphQL connectors to be _unnamed_.

In essence, this means omitting the `name` field in the directive.

For now, if a connector is unnamed, its name will be derived from the
internally tracked connector ID. Meaning, an unnamed connector will be
named `ConnectorN` where `N` is the internal ID number of the connector
(e.g. `Connector1`).

In the future, `name` will be renamed to `namespace`, and a connector
without a namespace will act accordingly (as in, no namespacing within a
separate root field, and types keep their upstream names).