fix: nested `@openapi` parameter serialisation (#2341)
We rename `@openapi` parameters in our GraphQL API. But weren't always
converting back to the OpenAPI names before sending data on to the
underlying API. This manifested itself in some weird ways, depending on
precisely what parameter & API you were using.

I've updated the `resolve_input` function to fix this.  

I've also introduced a flag that determines whether `resolve_input`
should apply these transforms or not. We've got customers that want to
use connector generated types in their resolvers, and passing data to
the resolver in a completely different format from what we present in
the API seems like it would be confusing. We'll probably want to think
about this when we get to a proper renaming/transforms RFC.