feat(auth): GB-3415: Implement public rule (#2323)
Implement public access. It is engaged when no auth header is sent, and
also in `private`/`group` access as a fallback.
This PR does not attempt to integrate the public rule with the owner
based auth as the latter works only with JWT-based auth (private/group
based).

On global level:
```
schema
  @auth(
    providers: [{ type: jwt, issuer: "https://idp.example.com", secret: "abc123" }]
    rules: [
      { allow: public, operations: [ read ] },
      { allow: private }
      ]
  ) {
  query: Query
}

```
On type(model) level:
```
type Todo @model @auth(rules: [{ allow: public, operations: [ read ] } ]) {
```
On field level:
```
id: ID! @auth(rules: [{ allow: public } ])
```
In order to use the field rules the public access needs to be first
enabled on the parent (type or globally).
Public field access is therefore enabled by default and there is no need
to make it explicit.