Gb 3588 retain editor state and history items (#1515)
To do:

- [x] get this working in nextjs

This PR introduces automatic state/history restoration, on load, to
Pathfinder.

The concept of history, in Pathfinder, differs from what you may have
come to expect from GraphiQL. In GraphiQL, history is a record of
operations that have been sent to the server...they do not include
response data. In Pathfinder, each operation within the document has an
execution/run history. We store all of this data in one object to make
setting and getting the _overall state_ of the document editor a
relatively simple exercise.

Pathfinder will remember the entire document in the document editor, the
cursor position (indicating which operation in the document is
"active"), and the execution history for each of the operations in the
document editor. To purge the history for an operation, simply remove
the operation from the document.

> This PR also fixes the pathfinder errors we've been seeing since last
week. `renovate` automatically updated `vitest-canvas-mock`, which uses
`jest-canvas-mock`, and `jest-canvas-mock` had an update that crushed
tests. Info
[here](https://github.com/wobsoriano/vitest-canvas-mock/issues/7).