chore: split structs out of MetaType (#2329)
GraphQL has 6 different kinds of types. But there's really no context
where a type can be any of the different kinds. There's usually a
smaller set of valid kinds for any given context - e.g. output fields
can be anything but `InputObjects`, input fields can only be enums,
scalars or input objects etc. etc.

In dynaql we model types with a single type, `MetaType`. This has struct
variants for each of the different kinds of type. This is great for
concise pattern matching, but makes it difficult to accurately model
GraphQL. You might know from context that a `MetaType` has to be an
`Object`, but you can't express that in rust types - you have to pass
around a `MetaType` and pattern match it whenever you want at its
contents.

This PR tries to fix this by splitting out the struct variants into
actual structs. This will let us pass instances of the inner structs
around, or even form alternative groupings (like `enum InputType {
InputObject, Scalar, Enum }`). It also makes it easier to use
convenience functions for building instances of `MetaType`, which I've
done in a bunch of places.

Sorry the PR is so big, but I couldn't see a meaningful way to split it
up.