fix: declare variables in `@graphql` queries (#2380)
If a user is using `@graphql` and uses a variable in a query, we might
need to pass that variable to the downstream server. We were correctly
passing the data to the downstream, but we forgot to declare the
variable in our query string so the downstream would error out.

This fixes that. I am not especially happy with the implementation - I
couldn't use half the methods in the existing `Serializer` because it
takes a `&mut self` and I ran into mutable/immutable borrow issues. But
the alternative seemed to be writing out another full traversal of the
query, which I didn't want to do right now.

I have some ideas for how we can make this kind of thing way easier in
the future, but they take time and who has that.