fix: re-case model names before using in relations (#2383)
We re-case model names into camelCase before using them, but we weren't
doing this before putting them into a MetaRelation.

This means that if a user had a type named `user` in their schema, then
the relation would be stored as `user` even though the model is called
`User`. At some point we use the model name from the MetaRelation, which
was causing it to be added to a list of `schemas_to_build` that some
query planning stuff uses. But when it went to build the schema it
failed, because it was lookign for `user`

This fixes that.